+++ internal/model/article.go
@@ -34,7 +34,7 @@ type ArticleGetListOutPut struct {
 	List  []ArticleListItem `json:"list" description:"列表"`
 	Page  int               `json:"page" description:"分页码"`
 	Size  int               `json:"size" description:"分页数量"`
-	Total int               `json:"total" description:"数据总数"`
+	Total int64             `json:"total" description:"数据总数"`
 }
 
 type ArticleSortMap struct {
