+++ internal/model/guestbook.go
@@ -16,7 +16,7 @@ type GuestbookGetListOutputItem struct {
 
 type GuestbookGetListOutput struct {
 	List  []*GuestbookGetListOutputItem
-	Page  int `json:"page" description:"分页码"`
-	Size  int `json:"size" description:"分页数量"`
-	Total int `json:"total" description:"数据总数"`
+	Page  int   `json:"page" description:"分页码"`
+	Size  int   `json:"size" description:"分页数量"`
+	Total int64 `json:"total" description:"数据总数"`
 }
